Put remote input into a separate thread.
The results I got from doing this were not as I expected
- I'm surprised I got this to work (and it was fairly easy)
- The remote control seems to have hiccups at start-up. I initially thought
this had to do with how I was handling input from the remote (which was poorly
done) and I thought putting the input traffic into a separate thread would fix
the issue. It did not. So now I'm thinking there might be an issue with the
streamer/music player finding information about the music directories at start-
up that I can't control. After about 30-40 seconds and punching some buttons
on the remote, the issue seems to be resolved.
- In separating the input data control to a separate thread, I was able to
set the boolean passed on music_context.iteration back to True, meaning
GStreamer has better control over the music-playback needs because it now makes
blocking calls. This also gets rid of the issue of consuming all the resources
on whatever core this process gets run on.