Came up with a way to reconnect to bluetooth
- So gstreamer has no issue connecting to a bluetooth device that isn't on
currently. When the device turns on, it connects almost immediately. However,
once the connection is lost (i.e. the bluetooth reciever/speakers turn off after
a certain amount of time not being used) the signal could not be recovered
- To detect the loss of signal, I set-up a handler to listen to the disconnect
signal on the OS's dbus. Once that gets triggered, the program hard-kills itself
with a special error code. This gets picked up by the wrapping shell script
and allows us to re-run the app from scratch. This, of course, is NOT a good
idea. BUT after a day's worth of fiddling with the getting consistent info
out of the dbus after the first disconnect and an inability to clean-up the
gstreamer pipeline it's the only way I could get the service to run reliably.